requirements.txt
fastapi
uvicorn
sqlalchemy
requests
google-generativeai
python-dotenv
pydantic
aiofiles
jinja2
